Fix file editor
