+++ package.json
@@ -24,7 +24,7 @@
   },
   "license": "GPL-3.0",
   "dependencies": {
-    "@markusjx/java": "0.0.8",
+    "@markusjx/java": "0.1.0-beta.2",
     "@wojtekmaj/react-daterange-picker": "^3.1.0",
     "custom-electron-titlebar": "^3.2.6",
     "electron-store": "^8.0.0",
